refactor: move utils that were composable to proper place
